+++ src/ldc/translation/_jsonlines.py
@@ -108,7 +108,7 @@ class JsonLinesTranslationReader(TranslationReader):
         self._current_input = self._inputs.pop(0)
         self.session.current_input = self._current_input
         self.logger().info("Reading from: " + str(self.session.current_input))
-        self._current_input = open_file(self._current_input, mode="rt", encoding=self.encoding)
+        self._current_input = open_file(self._current_input, mode="rt", encoding=self.encoding, logger=self.logger())
 
         self._reader = jsonlines.Reader(self._current_input)
         for item in self._reader: